drm: always try enabling universal plane support
Always enable the DRM universal planes capability, in order to receive
information about all planes from DRM/KMS, and not only from overlay
ones. Previously this was set only when the cursor support is enabled,
but it is needed also to modify properties of primary planes e.g. to
set rotation on them.