drm/gles: Perform the GBM/EGL initialization dance
Add the code needed to allocate and prepare a GBM surface and its
corresponding EGLSurface. Note that the EGL configuration attributes
specify a particularly lax set of restrictions in order to let the
driver return its most preferred (and typically faster) one, and
then use eglGetConfigAttrib() to fetch the actual pixel format it
has chosen to pass it down to gbm_surface_create(). This is much
nicer than hardcoding values, and should be more robust as well.