drm, cog: Allow passing parameters at platform setup time
Add a new --platform-params/-O command line option and have its value
passed down to cog_platform_setup() to allow passing an arbitrary string
to the platform module initialization. While at it, allow setting both
the rotation and renderer properties of CogDrmPlatform this way, which
is in particular important for the renderer property because it is
construct-only and there is no suitable mechanism to set construct-only
properties for CogPlatform subclasses at the moment.