core: Remove unneeded CogPlatformClass.teardown vfunc
Use GObjectClass.finalize instead of CogPlatformClass.teardown. Now that
CogPlatform derives from GObject, it is possible to reuse its functionality
instead of having an ad-hoc vfunc that gets called immediately before
destroying an instance. As a bonus, not only the implementations follow
a more idiomatic GObject approach, but also it is not possible anymore
to forget a call to cog_platform_teardown() before destruction.