core: Move CogLauncher from the library into launcher
Move the CogLauncher class implementation from core/ to launcher/, as in
practice the way it works is tightly tied to the "cog" program and quite
inflexible, preventing its reuse in other programs. As a matter of fact,
third party launcher that use libcore typically reach out for CogShell
and the platform plug-in loading capabilities instead. All in all, it
seems like a better idea to convert it into an implementation detail of
the "cog" launcher.

The only drawback is that the Wayland platform plug-in relied on being
able to fetch the current GApplication and web the view. The actions
taken on key press (go back, forward, change zoom, exit application) are
likely undesired in embedded scenarios, so skip them if the is not an
active GApplication.