CI: Make curl downloads more reliable
Retry the curl command up to ten times in case of download failures.
This should alleviate build issues like the following, where the server
sometimes seems to wrongly close HTTP/2 streams, or when some connection
issue results in partial downloads:

    https://github.com/Igalia/cog/runs/4176134469
    https://github.com/Igalia/cog/runs/4176255670

While at it, pass "-C -" to let curl determine from which point to retry
the download. As the toolchain packages are big, this should help avoid
timeouts.

Finally, use HTTP/1.1; because using HTTP/2 seems to be a source of
sporadic issues, and for a big, bulky download, the gains from HTTP/2
(like multiple streams and faster connection open) are basically none
due to the transfer itself taking most of the time.