launcher: Place built binaries in top level directory
Explicitly place the built "cog" and "cogctl" binaries in the top level
build directory. The location was moved in #376 without giving a second
thought about it but it broke the "run-minibrowser" script from the
WebKit repository, see: https://bugs.webkit.org/show_bug.cgi?id=234132

On top of preveting breakage of third party tooling, setting the build
directory this way is more convenient for running programs for testing
during development.