core: Declare public symbol visibilities
Change the build options to use "hidden" visibility by default, and mark
only the symbols that are part of the public API with the COG_API macro.
This prevents library private symbols from leaking to the global symbol
namespace even if they are not declared as "static", and also makes the
ELF PLT smaller which prevents accidental symbol overriding and improves
load times.