wl: Check seat interface version properly
Use accessors to get the version of the protocol at runtime, instead
of using wl_seat_interface.version, which is a compile-time constant
which depends on the Wayland version used for the build, and does not
reflect which version of the protocol is being used at run time.

Instead of assuming that pointer frame events are available when the
Wayland version used at build time is 1.10 or newer, do a feature check
on WL_POINTER_FRAME_SINCE_VERSION instead. This can be used also at run
time, which covers the case in which a compositor may not support the
newest version of the protocol but the code can handle gracefully by
checking the version at run time with wl_pointer_get_version().

While at it, arrange the requested version of the wl_seat protocol when
binding the registry global to pick the most recent version supported
by the compositor inside the [3, 7] range.

(cherry picked from commit 0406b2195398bfce7c205426e5708c41290a3d5a)