x11: Support setting view activity state flags
Implement setting view activity state flags, which allows WebKit to
perform certain optimizations (like skipping updates while the view
is not present in a window, or not visible) and makes certain features
work as expected (like the Page Visibility API).

- Events XCB_FOCUS_{IN,OUT} modify wpe_view_activity_state_focused.
- Events XCB_{UN,}MAP_NOTIFY modify wpe_view_activity_state_in_window.
- Event XCB_VISIBILITY_NOTIFY modifies wpe_view_activity_state_visible;
  both partially obscured and completely unobscured windows are set
  as visible.

The corresponding event masks are added during window creation.

Interestingly, at least on GNOME Shell in an X11 session, Cog does not
get any XCB_VISIBILITY_NOTIFY events after the initial one, even if the
window is hidden iconified. This is likely caused by the compositor
wanting to have windows update so it can have an always a fresh view
of the contents e.g. to paint window miniatures in a task switcher or
in overview mode.