x11: Fix glitches in non-composited window managers
Handle X11 expose events, which was missing. Typically composited
environments (running xcompmgr, a composited desktop like GNOME Shell,
or similar) never produces expose events: the compositor will keep
around a buffer with the latest contents of the X11 windows, in order to
perform composition, which means that it does not ever need to request
application to repaint windows. Given that composited environment are
common these days, that is likely the reason why this issue has not been
noticed earlier.

On the other hand, with more "traditional", non-composited window
managers will happily discard pixmaps for obscured windows (in order
to save memory, for example) and rely on the fact that X11 applications
are expected to repaint themselves when requested via expose events. In
order to be able to do this, the X11 platform plug-in now keeps the
latest frame exported by WPEBackend-fdo, and performs a repaint when an
expose event indicates that the window (or part of it) is visible.

The existing "needs_initial_repaint" flag is renamed as "needs_repaint"
and reused also for expose events. Repaints are scheduled by sending
a notice event, similarly to how frame completions are postponed queued
to the event loop.

This is a follow-up to #398