build: Homogenize linked wpebackend-fdo version
First of all bump cmake minimum required version to 3.12 to sync with WebKit
requirements and use some features for this changeset.

Second, the patch adds a FindWpeFDO.cmake module to find a common wpebackend-fdo
library for all the compiled platforms; otherwise, there might be linking
poisoning at compilation time, if a custom wpebackend-fdo is used (by the
uninstalled method, for example) different to the one system-wide.