Fix COG_WESTON_DIRECT_DISPLAY build
For some reason this option was enabling content protection as well,
since that is a seprate option that merely depends on direct display
move it under COG_WESTON_CONTENT_PROTECTION.

We also need to fix pkg-config search to look for libweston protocols
starting with weston version 8(first to introduce direct display),
the current weston development branch uses version 11 so lets include
that as well.

We also need to fix the client protocol header names as the current
ones are invalid.

Signed-off-by: James Hilliard <james.hilliard1@gmail.com>