x11: handle upper case letter, caps lock, num lock, control and alt
Fixes #391

Steps to test:
1. Run COG_MODULEDIR=$(pwd)/modules ./cog -P x11 'data:text/html,<input>'
2. Enter various key combinations (with shift, control, caps lock, num lock, etc), and verify that they work.

What was missing was updating the xkb state on each key press.
That needs to be done before the call to xkb_state_key_get_one_sym
as xkb reads the state in order to determine which keysym it should
output.

The function is more complex that in need to be because XCB, X11,
and WPE all use different masks for the modifiers.

Future improvements to this code might include:
* unifying xcb_handle_key_press and xcb_handle_key_release
* handle xkb keyboard map changes
* handle AltGr