gamepad: Select backend at launch time.
Adds the command line option `--gamepad` where the user can select, at launch
time, the backend of the gamepad to use. By default `disable` is available. If
cog is compiled with libmanette as dependecy, also `manette` backend would be
available.

The initialization of the gamepad is split in two stages:

1. When the launcher selects the backend.
2. When the platform sets the `get_view_backend()` callback.