cmake: Write correct wpe-webkit requirement in .pc file
Arrange for the generated cogcore.pc file to include the version
of the wpe-webkit-1.x dependency, instead of hardcoding 1.0; this
solves issues when Cog was configured against wpe-webkit-1.1 where
the .pc file would have the wrong version.

The switch from Requires.private to plain Requires is needed to
let pkg-config add the WPE WebKit library in the output for
"pkg-config cogcore --libs", otherwise users need to specify
both cogcore and wpe-webkit-1.x as dependencies, when using
cogcore should be enough.