wl: Do not attach frames with invalid size to the surface
Skip frames which do not match the size of the output surface, in order to
avoid a Wayland protocol error resulting from mismatched sizes. This can
happen, for example, after requesting that the surface is maximized, while
the view backend has already produced some frames, but the change of size
has not been yet noted by WebKit. In these cases we can safely skip using
the frames but they still need to be ACK'd in order to let WebKit continue
producing new ones, which eventually will arrive in the desired size.

Related issue #537