wl: Add overview and troubleshooting section in docs
Improve the Wayland platform plug-in documentation by including an
overview that mentions the need to run a Wayland compositor, lists
some known-to-work compositors, and clarifies the need to have the
environment variables set properly. As the cherry on top, include
a troubleshooting section with a check list to go through and tips
for figuring out the values for environment variables.