headless: Cleanup in preparation for multiple view support
Tidy up the headless backend in preparation for adding multiple view
support down the line. In particular, rename "struct platform_window"
to CogHeadlessView, and embed an instance in the CogPlatform struct
for now, removing the "win" global.

While at it, remove the "view_backend" member from CogHeadlessView,
because once it is returned from the get_view_backend vfunc the
ownership is transferred anyway, and it is not needed elsewhere.

Rename setup_fdo_exportable to cog_headless_view_initialize, and leave
out the wpe-fdo initialization in cog_headless_platform_setup. Add a
matching cog_headless_view_finalize, which takes care of disarming the
tick timer. Destroying the view backend is now done using the
destroy_notify callback from WebKitWebViewBackend, which is more
correct.