CI: Avoid duplicated jobs on pull requests
Restrict push jobs to development branch and release branches, to avoid
having a duplicate set of CI jobs when filing a PR associated to the
push event to the branch from which the PR was created. Also, restrict
pull_request jobs to those PRs targeting the same set of branches.