meson: Bring back support for using wpe-webkit-1.1
Bring back support for the WPE WebKit 1.1 API, which Cog should support
as long as WebKit does. This removes the soup2 Meson option in favor of
a new wpe_api option, which chooses both the API level we are targeting
and decides on which version of libsoup to require depending on that.
This works because the WebKit API implies which libsoup API level is
being used.