CI: Save checksum of toolchain installer insted of timestamp
Improve the logic around reusing a cached toolchain installation by
storing the toolchain installer checksum file instead of the time of
installation. This way we can check that the cached installation
corresponds to what is expected, or do it afresh otherwise. This is
more robust than the existing approach that trusts that any file on
the server newer than the installation timestamp should be preferred.