shell: Replace single view with CogViewStack group
Use a CogViewStack instead of a single web view. This will allow a shell
to display one of a number of web views in the output for the platform
plug-in being used. While some plug-ins may support multiple outputs
(e.g. windows, Wayland surfaces, DRM/KMS planes, etc.) they need to be
cleaned up to avoid using globals, but in the meantime it is possible
to use the same output to display one single view of may, which is
exactly what CogViewStack helps with.