core: Add a fallback CogPlatform implementation
This adds a new CogFallbackPlatform implementation, which gets
built into the core library. Instead of requiring applications to
implement themselves the needed code to use plain WPE backends (like
WPEBackend-rdk), they can rely on a CogPlatform instance being always
instantiated and available. The implementation is registered to the
GIO extension point automatically and always available.

This new platform has the lowest possible priority and needs to be
manually selected: while it is debatable that it should have more
priority than the headless platform, keeping it this aligns better
with existing behaviour.