wl: Support multiple web views
Introduce a CogWlView class, allowing for multiple web view support.
The view class tracks the FDO backend exportable, the last received
buffer export, and a few other details needed for updating the view,
and for showing it back when hidden and becoming visible again. Moving
those allows for the removal of the "wpe_host_data" and "wpe_view_data"
globals.

Most of the changes involved arranging to pass around either the
CogWlPlatform instance of the CogWlView as user data to callbacks,
plus replacing the single view with the CogViewStack used by the shell
taking care that only the visible view of the stack gets to attach
buffers to the Wayland surface and that input events are sent to the
visible view.

Note that this disables SHM buffer exports, input methods, and gamepad
support, which would be re-enabled later on. Some of these features use
hardcoded globals and need to be refactored to avoid them.