core: Move key binding handling into CogView
Pick the key binding handling from the Wayland platform plug-in, and
move it into the core library. A new cog_view_handle_key_event() method
takes care of detecting supported key bindings or actually sending the
events to the view backend otherwise. In order to support cases where
key bindings are not desirable (e.g. kiosks) it is possible to disable
key binding support.

The only bit missing to move is the F11 key binding to (un)fullscreen
as the operation may need platform view specific actions, so for now
the handling for this binding is left as it was in the Wayland platform
plug-in. This can be improved later on in a follow-up patch.