wl: Handle pointer and touch events at the window level
Move handling of pointer and touch events from CogWlDisplay to
CogWlWindow. In Wayland events are always local to a surface i.e. (x, y)
coordinates relate to a certain surface (in our case: a window) and not
to the display itself.

For this change the CogWl{Pointer,Touch} event is moved into
CogWlWindow, and the .enter callbacks are used to track the surface
(that is: window) to which subsequent events belong. The CogWlWindow is
set as user data for their corresponding Wayland surfaces to easily
track the surface->window relationship.