wl: Force dispaching a frame complete when the View is set visible
By dispatching this frame complete event we are forcing to the backend
to generate a new frame. This is needed for those views who have been
loaded on a hidden wl_surface. Under those conditions the frame done
callback is not delivery from the compositor since the frames are not
being released by it (not painted but waiting in case the surface
becomes visible).