wl: Add an XdpParent on file chooser dialog
The XdpParent needs to know the "parent_window" string to keep the dialog on top.
On wayland this "parent_window" can be obtained using the xdg-foreign protocol.
This protocol is used on this XdpParent implementation.
First it must call export_toplevel method, and then register a callback to
receive a handle which is the "parent_window".

To store the required data for this XdpParent implementation it is used the
private "data" pointer property as a xdp_parent_wl_data.
XdpParent isn't responsible of freeing it, so that data is handled at
cog-platform-wl.