wl: Create custom CogWlView
No logical changes in the workflow. This change only creates the
CogWlView as a specialization of the CogView. We also move the logic
for handling the exported buffers inside of this new class.

This also brings other benefits since this also removes the access to the
global variables by adding context data to the listeners defined in the
CogWlPlatform and in the CogWlView classes. For this reason, for example,
the s_display is not longer needed as global and now is moved and
accessible as part of the CogWlPlatform.

The are additional cosmetic changes like these:

* The struct win_data is renames as CogWlWindow
* Some functions are renamed with the cog_wl_* and cog_wl_view_* prefix

This change is motivated towards the progress of being able to have more
than one view.