wl: Support multiple wl_seats
A CogWlDisplay contains a default seat. This default value is the first seat
registered and hereafter the last one activated. This is used for the
shell_popup_surface and for the on_mouse_target_changed handlers. The
CogWlDisplay also contains the rest of the CogWlPlatform in a list.

This patch also fixes the tracking of the serials for the input methods.
Now each different kind of input method for each Seat is stored
individually and updated according. Still it is possible to get the
latest serial using the suitable cog_wl_seat_get_serial() function.

Some code was cleaned-up and relocated in cot-utils-wl as
functions: cog_wl_display_add_seat; cog_wl_seat_create();
cog_wl_seat_destroy(); void cog_wl_text_input_clear(); cog_wl_text_input_set()