wl: Add CogWlPlatform::resized-window signal
The `resized-window` signal is emitted when the window is resized.

Handling this signal allows to react to changes in the geometry of
the window and decouples the CogWlView from the CogWlPlatform internal
logic.

The view (or the views in the future) is now connected to this signal
and reacts according.

This change is motivated towards the progress of being able to have more
than one view.