wl: Remove CogWlPlatform::init_web_view()
All the initialization done in the init_web_view() is now relocated
inside of the cog_wl_view_init(). To make this possible all the code
related to the "run-file-chooser" and "mouse-target-changed" signals
have be moved from the cog-platform-wl to the cog-view-wl.

This change makes the Wayland platform compatible with the check
implemented in the cog_platform_init_web_view(): "CogWlPlatform class
defines both .get_view_type and .init_web_view the latter should be
removed."