wl: Refactor popup_data as CogWlPopup
There is not longer a static global popup_data definition inside of
the platform. Instead of this the Platform will have attached a
CogWlPopup instance.

It is the handling of the CogWlPopup responsibility of the
CogWlPlatform object. For example, the cog_wl_platform_popup_create()
and the cog_wl_platform_popup_destroy() will deal with the set and
unset of the reference to the popup object inside the CogWlPlatform.

Because this reason, the CogWlView just relies on the CogWlPlatform
for the creation and the rest of the life-cycle of a CogWlPopup now.
All this is simply triggered by the on_show_option_menu() handler.

There are also other trivial refactorings:

* The cog_wl_compositor_create_surface() is moved to the cog-utils-wl
  for convinience.
* The popup auxiliary structs and callback handlers are also moved to
  the cog-utils-wl.
* The CogWlPopup is now accesible from the CogWlPlatform object but we
  still try to avoid global access to the platform by passing the
  platform object, as much as possible, as contextual information to
  the handlers.

Reviewed-by: Adrian Perez <aperezdc@igalia.com>