wl: Fix the validation of the exported image in outputs with scale factor
The patch fixes a basic logic bug inside of the on_export_wl_egl_image().
The error is in the incorrect use of scale_factor (actually defined for the
WebPage zooming) instead of the current_output->scale for validating the
exported image geometry.

The change also unifies inside of the new validate_exported_geometry()
function the shared code used for validating the sizes of the exported images.

This fixes regression added in e533fce.

Thanks to Adrian Perez de Castro <aperez@igalia.com> for reporting the issue.