wl: Support multiple web views
Replace the single view saved in the CogWlPlatform by the viewport
being used by the shell, taking care that only the visible view gets
to update the contents of the Wayland surface and that input events
get sent to the visible view as well.

There are some locations where the visible view is assumed to be also
the focused one, which may not be true later on when multiple viewports
are supported. This compromise is okay for the time being and may need
to be revisited later on.