wl: Shell handles many Viewport
Added new functions in the CogShell API for handling multiple Viewport:

* cog_shell_viewport_lookup(CogShell *, CogViewportKey);
* cog_shell_viewport_new(CogShell *, CogViewportKey);
* cog_shell_viewport_remove(CogShell *, CogViewportKey);

Also added the CogShell::create-viewport signal. This signal is
emitted when the shell creates a new a new CogViewport. Handling this
signal allows to customize the actions required to be done during the
creation of a new viewport.

This change also modifies the behaviour of the 'cog_shell_get_viewport()'.
It  creates a default viewport in order to keep the current behavior. The
creation of the default viewport is lazy and it is delayed until the
invokation of the 'cog_shell_get_viewport()'.