wl: Shell handles many viewports
Add new functions in the CogShell API for handling multiple Viewports:

* cog_shell_viewport_lookup(CogShell *, CogViewportKey);
* cog_shell_viewport_new(CogShell *, CogViewportKey);
* cog_shell_viewport_remove(CogShell *, CogViewportKey);

Also added the CogShell::create-viewport signal. This signal is
emitted when the shell creates a new CogViewport. With handling this
signal is possible to customize the required actions to be done during
the creation of a new viewport.

This change modifies the 'cog_shell_get_viewport()'. Now it creates a
default viewport in order to keep the current behavior. The creation
of the default viewport is lazy and it is delayed until the invocation
of the 'cog_shell_get_viewport()'.