wl: Shell handles many Viewport
Added new functons in the CogShell API for handling multiple Viewport:

* cog_shell_viewport_lookup(CogShell *, CogViewportKey);
* cog_shell_viewport_new(CogShell *, CogViewportKey);
* cog_shell_viewport_remove(CogShell *, CogViewportKey);

Also added the CogShell::create-viewport signal. This signal is
emitted when the shell creates a new a new CogViewport. Handling this
signal allows to customize the actions required to be done during the
creation of a new viewport.

This modifies the behaviour of the 'cog_shell_get_viewport()'. This
creates a default viewport in order to keep the current behavior. This
creation of the default viewport is lazy and it is delayed until the
invokation of the 'cog_shell_get_viewport()'.

Finally, this also clarifies how the launcher based on Cog should to
perform the start-up:

1. Create the Platform
2. Create the Shell
3. Configure the platform using the created Shell
4. Shell startup

(see documentation in 'cog_launcher_startup()').