wl: Check wl_surface before use it
... on pointer_on_enter, touch_on_down and keyboard_on_enter.

It can be possible a race condition during the destruction of a CogView
if, during this action, the UI-process is receiving and trying to
process an input event.

Under this scenario, the wl_surface associated could be not longer
there so the input handler will receive a null-pointer (0x0).

Backtrace:

```
0  0x6f45a3bc in wl_proxy_get_user_data (proxy=proxy@entry=0x0)
    at ../wayland/src/wayland-client.c:2135

    at /usr/include/wayland-client-protocol.h:3393
    fixed_x=0, fixed_y=5, surface=0x0, data=0x6cf09088) <<<<<<<<<<<<<
```