wl: Check wl_surface before use it
... on pointer_on_enter, touch_on_down and keyboard_on_enter.

It can be possible a situation during the destruction of a CogView
where the UI-process were receiving and trying to process an input
event for a non existant wl_surface.

Under this scenario, the wl_surface associated could be not longer
accesible so the input handler will receive a null-pointer (0x0).

Backtrace:

```
0  0xb9999999 in wl_proxy_get_user_data (proxy=proxy@entry=0x0)
    at ../wayland/src/wayland-client.c:2135

    at /usr/include/wayland-client-protocol.h:3393
    fixed_x=0, fixed_y=5, surface=0x0 <<<< , data=0xa9999999)
```