core: Remove cog_platform_configure() function
Remove the cog_platform_configure() utility function, because at this
point it does not do much else than calling cog_platform_new() and
cog_platform_setup(). Instead, make sure that the platform instance
is a global singleton and that it is created on-demand on the first
call of a the cog_platform_get() accessor.

In order to allow manually indicating which platform plug-in to use, a
new cog_init() function is added. While at it, make it unnecessary for
programs to always call cog_modules_add_directory() if they only need to
use the default module path.

The upside of these changes is that the API gets much more convenient
to use, and in the simplest of cases, it is enough to let the library
initialize the platform instance. The one-time configuration using
cog_platform_setup() is still needed, but now the following is enough
to properly use the library:

  CogShell *shell = cog_shell_new("my-program", FALSE);
  if (!cog_platform_setup(cog_platform_get(), shell, "", NULL))
      g_error("Cannot initialize Cog.");

Note that the COG_MODULEDIR and COG_PLATFORM_NAME environment variables
are still used when NULLs are passed to cog_init() or during implicit
initialization.