core: Allow passing NULL as platform setup params
Change cog_platform_setup() to allow passing NULL as the parameters
string, and fall back to using the value of the COG_PLATFORM_PARAMS
environment variable in that case. In order to keep the same ABI with
existing platform plug-ins, the function ensures that an empty string
is always passed when needed.

This change simplifies call sites for the function, which no longer
need to handle reading the environment variable themselves, nor taking
care of always passing a valid string, even if empty.

While at it, write the missing documentation for cog_platform_setup().