platform: Clean up multiple cursor_names definition
Move the cursor names from the header to a C source, provide a helper
function to get the list of names for a cursor type, and while at it
make the naming of types and functions be idiomatic.

The cursor_names global would be defined every time the cursors.h header
was included, which is not optimal. Also, including the header and not
using the variable caused compiler warnings.