core: Make the CogView.viewport property only readable
Remove the G_PARAM_CONSTRUCT flag from the CogView.viewport property,
because it it supposed to be only readable and it the viewport is
set in a separate step separate from view construction by using
cog_viewport_add(). It would be technically possible to allow setting
the viewport at construction time, but that would need some additional
support code it would use G_PARAM_CONSTRUCT_ONLY. For now it is easier
to remove the flag.