wl: Fix the COG_PLATFORM_WL_VIEW_FULLSCREEN not works
The COG_PLATFORM_WL_VIEW_FULLSCREEN is evaluated during inside of the
cog_wl_viewport_create_window(). At that point there is not yet a view
created in the viewport associated to the window.

Allow the shells to set the fullscreen mode even if there is not a view
available in the viewport. This will make the new views being created
in fullscreen mode. This applies also to the one created during the
initialization.