x11: Support using xcb-keysyms as fallback for XKB
Massage the parts from #303 which deal with input using xcb-keysyms to
be included in the x11 platform plug-in. This provides a fallback that
can be used when XKB initialization fails, typically in cases where the
X server does not support XKB or has the support disabled. XKB is a X11
protocol extension after all, and it is wrong to assume that it will be
always available. While at it, allow running without keyboard support,
producing a warning in this case.