meson: Add a Wrap subproject for xcb-keysyms
Add a Wrap subproject definition for the xcb-keysyms dependency. It is
a very simple library with a single C source file, so it is easy to make
Meson build definitions for it that get added on top of the tarball
release fetched by Meson if xcb-keysyms is not available. This way it is
guaranteed that there will be at least some support for keyboard input
if the XKB extension is not available in the X server, or not enabled at
build time.