automation: Save automated view and avoid dangling ref when closing window
The view created for automation was not being saved in the viewport,
leaving some required signals unhandled, like "close". This causes the
view to never be destroyed when called from WebKit's
WebPageProxy::close(), leading to infinite loop in the close window
WebDriver algorithm, as the list of open windows would never change.

Also, as the WebKitAutomationSession::create-web-view doesn't transfer
the ownership of the returned view, we make the viewport hold the single
initial reference to the newly created view.

This should fix the WPT webdriver fixtures that rely on creating and
destroying views for testing.

Note: This commit makes creating new windows for automation behave like
creating new tabs, regardless of the "create-web-view::window" detail,
which maybe could be implemented creating a new viewport in a future
commit.