meson: Use a generated depfile for gi-docgen
Convert the output from "gi-docgen gen-deps", which is a list of files
ingested to generate the documentation, into a format suitable to be
read by Ninja (a Make snippet), and use it as the "depfile" for the
documentation custom target. The transformation is done using a small
Python script, which does not introduce any new dependency because
gi-docgen itself already required Python to be present.